Update to compile on MacOS Tahoe 26.3 with latest packages on MacPorts. CMakeList.txt was updated to use the modern way to find Boost (see policy CMP0167). Argument to boost::posix_time::microseconds is now explicitly cast as a long to avoid a compilation error.

+++ src/braille/braillegen.cpp
@@ -266,7 +266,7 @@ void BrailleGen::SetVibFreq(float hz)
 {
 	LATERO_GRAPHICS_GUARD
 	SetLastModified_();
-	vibPeriod_ = boost::posix_time::microseconds(1E6/hz);
+	vibPeriod_ = boost::posix_time::microseconds((long)(1E6/hz));
 }
 
 float BrailleGen::GetVibFreq() const
